move navbar, footer constants to respective files
