+++ src/contexts/globalLayout/GlobalFooter/GlobalFooter.tsx
@@ -3,7 +3,6 @@ import breakpoints, { pageGutter } from 'components/core/breakpoints';
 import { BaseS } from 'components/core/Text/Text';
 import Spacer from 'components/core/Spacer/Spacer';
 import colors from 'components/core/colors';
-import { GLOBAL_FOOTER_HEIGHT, GLOBAL_FOOTER_HEIGHT_MOBILE } from '../constants';
 import {
   GALLERY_FAQ,
   GALLERY_JOBS,
@@ -67,6 +66,9 @@ type StyledFooterProps = {
   isMobile: boolean;
 };
 
+export const GLOBAL_FOOTER_HEIGHT = 80;
+export const GLOBAL_FOOTER_HEIGHT_MOBILE = 134;
+
 const StyledGlobalFooter = styled.div<StyledFooterProps>`
   display: flex;
   justify-content: space-between;
@@ -78,6 +80,7 @@ const StyledGlobalFooter = styled.div<StyledFooterProps>`
 
   background-color: ${colors.white};
 
+  // TODO: all of this can go away with new layout + NFT Detail Modal update
   @media only screen and ${breakpoints.tablet} {
     padding: 0 ${pageGutter.tablet}px 24px;
