--- src/contexts/globalLayout/constants.ts
@@ -1,5 +1,0 @@
-// Note: These values are the same on purpose! Changing this will require
-// re-thinking how the app centers content between nav + footer
-export const GLOBAL_NAVBAR_HEIGHT = 72;
-export const GLOBAL_FOOTER_HEIGHT = 80;
-export const GLOBAL_FOOTER_HEIGHT_MOBILE = 134;
