+++ src/scenes/NftDetailPage/NftDetailAsset.tsx
@@ -1,6 +1,5 @@
 import breakpoints, { size } from 'components/core/breakpoints';
 import styled from 'styled-components';
-import { GLOBAL_FOOTER_HEIGHT, GLOBAL_NAVBAR_HEIGHT } from 'contexts/globalLayout/constants';
 import NftDetailAnimation from './NftDetailAnimation';
 import NftDetailVideo from './NftDetailVideo';
 import NftDetailAudio from './NftDetailAudio';
@@ -13,6 +12,8 @@ import NftDetailImage from './NftDetailImage';
 import NftDetailModel from './NftDetailModel';
 import { useMemo } from 'react';
 import { getBackgroundColorOverrideForContract } from 'utils/nft';
+import { GLOBAL_FOOTER_HEIGHT } from 'contexts/globalLayout/GlobalFooter/GlobalFooter';
+import { GLOBAL_NAVBAR_HEIGHT } from 'contexts/globalLayout/GlobalNavbar/GlobalNavbar';
 
 type NftDetailAssetComponentProps = {
   nftRef: NftDetailAssetComponentFragment$key;
